E-learning-chatbot/requirements.txt
streamlit>=1.32.0
google-generativeai
ollama
chromadb
langchain
langchain-community
langchain-google-genai
sentence-transformers
pypdf
python-dotenv
tiktoken